refactor: 요금 계산 로직을 Path로 이동
